# WIP: Explore using Aho-Corasick for specific patterns in Regex