This is draft PR, strictly for prototyping and exploring using Aho-Corasick (AC) for specific patterns in Regex and to gather initial feedback.
cc @danmosemsft @jeffhandley @eerhardt @stephentoub 
Prelim perf results:
```
Benchmark:
        _ahoCorasickMultipleWords = new Regex("(efg|xyz|hij)abcd");

        [Benchmark] public void AhoCorasickMultipleWords() => _ahoCorasickMultipleWords.IsMatch(@"hhhhhhhhhhhhijabcd");

```
```
D:\repos\performance\src\tools\ResultsComparer>dotnet run --base "D:\repos\before_aho" --diff "D:\repos\after_aho" --threshold 0.001%
summary:
better: 3, geomean: 3.047
total diff: 3

No Slower results for the provided threshold = 0.001% and noise filter = 0.3ns.

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.RegularExpressions.Tests.Perf_Regex_Common.AhoCorasickMultipleWords( |      3.09 |          1800.19 |           583.00 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Common.AhoCorasickMultipleWords( |      3.07 |          1775.07 |           578.28 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Common.AhoCorasickMultipleWords( |      2.99 |          1762.12 |           590.22 |         |

```

Over the past couple weeks, I spent some time prototyping AC in Regex. As https://github.com/dotnet/runtime/issues/1349 points out, we should expect to see perf gains when alternations are present in a Regex. 

Current code and Prototype:
We have a `FindFirstChar` and `Go` loop to find matches in a given input. `FindFirstChar` looks for the first matching character in the input. Once we find a matching character, `Go` runs a loop to see if the following characters match a word in the Regex. This logic is however not efficient for multiple words. For ex: If we have the pattern `(efg|xyz|hij)abcd`, `FindFirstChar` looks for either `e, x, h`. Then `Go` checks for `efgabcd, xyzabcd and hijabcd` in that order. So, for the input `hhhhhhhhhhhhijabcd` (12 `h` followed by abcd), at each `h`, `Go` performs 3 checks => 11 * 3 = 33 checks before the 12th `h` is matched and `hijabcd` is found. In contrast, the AC algorithm scans linearly, so we perform only 11 checks before the 12th `h` is matched and we find `hijabcd`. This is where the speed up comes from!

Note: The benchmarks lie in the sense that the setup phase of AC is not accounted for. The allocations to prepare `_ahoCorasickWords` are not measured in the benchmark. Also, the pattern and text used are the specifically chosen to see the speedup, but they are valid nonetheless :)