@stephentoub 2020-12-07
Thanks for working on this.  I've not yet looked at your changes (and probably won't be able to until January), but you might try running the regex redux benchmark,  as it has a bunch of alternations.