@danmosemsft 2020-12-07
I'm interested in regexredux too, I guess you will need to convert `ProcessString(..)` to take a span so it can do multiple matches.