@pgovind 2020-12-16
> interested in regexredux too

> running the regex redux benchmark

Yup, I do plan on running the regex redux benchmark once this prototype is fully ready. Will post the results here then. Thanks for the reviews here meanwhile!